requirements.txt
arrow==1.2.3
asgiref==3.5.2
attrs==21.4.0
binaryornot==0.4.4
certifi==2022.5.18.1
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==2.1.1
crispy-tailwind==0.5.0
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.0.4
django-allauth==0.51.0
django-allauth-ui==0.1.0
django-appconf==1.0.5
django-browser-reload==1.6.0
django-compressor==4.3.1
django-crispy-forms==1.14.0
django-generic-positions==0.2.3
django-hvad==1.8.0
django-libs==2.0.3
django-rated-reviews==1.0.1
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
easy-thumbnails==2.8.5
Faker==12.0.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
jinja2-time==0.2.0
lark-parser==0.12.0
MarkupSafe==2.1.2
mixer==7.2.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-slugify==8.0.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
rcssmin==1.1.1
requests==2.28.0
requests-oauthlib==1.3.1
required==0.4.0
rjsmin==1.2.1
sentry-sdk==1.15.0
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.14
